remove r_executor.py and add in RSerializeExecutor with comments
